Ignore unknown attribute assigns on models
Rather than raising an error when a model is assigned an unknown
attribute, silently assign it and ignore. This will prevent changes
to the Canvas API docs from causing errors in the Pandarus client
when new attributes are introduced on the server side.

Provides a little more detail on TypeError exceptions.

Also update to latest Canvas API docs.

Bump to 0.4.5

Fixes SIS-308

Change-Id: I05b9c00c1b1dd51e23f99335939739509ca0e20f